coinnews, bitwindow: extract codec into its own Go module
