+++ coinnews/codec/codec_test.go
