+++ coinnews/codec/constants.go
