+++ coinnews/codec/decode.go
