+++ coinnews/codec/doc.go
