+++ coinnews/codec/encode.go
