+++ coinnews/codec/go.mod
@@ -1,0 +1,17 @@
+module github.com/LayerTwo-Labs/sidesail/coinnews/codec
+
+go 1.25.0
+
+require (
+	github.com/btcsuite/btcd/btcec/v2 v2.3.6
+	github.com/stretchr/testify v1.11.1
+)
+
+require (
+	github.com/btcsuite/btcd/chaincfg/chainhash v1.1.0 // indirect
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+	github.com/decred/dcrd/crypto/blake256 v1.1.0 // indirect
+	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.4.1 // indirect
+	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
