+++ coinnews/codec/io.go
