+++ coinnews/codec/message.go
