+++ coinnews/codec/sign.go
